# server/docs/swagger.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "TO-DO list app",
		"description": "Endpoints for TO DO list application.",
		"version": "1.0.0"
	},
	"paths": {
		"/api/signup": {
			"post": {
				"summary": "Create user account",
				"tags": ["Authentication"],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"username": {
										"type": "string",
										"example": "john_doe"
									},
									"password": {
										"type": "string",
										"example": "securePassword123"
									}
								},
								"required": ["username", "password"]
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "User created successfully",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"example": "User created successfully"
										}
									}
								}
							}
						}
					},
					"409": {
						"description": "User already exists",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "string",
											"example": "User with this username already exists"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Validation error"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			}
		},
		"/api/login": {
			"post": {
				"summary": "Login user",
				"tags": ["Authentication"],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"username": {
										"type": "string",
										"example": "john_doe"
									},
									"password": {
										"type": "string",
										"example": "securePassword123"
									}
								},
								"required": ["username", "password"]
							}
						}
					}
				},
				"responses": {
					"204": {
						"description": "Login successful"
					},
					"404": {
						"description": "User not found",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "string",
											"example": "User not found"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Invalid password",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "string",
											"example": "Invalid password"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Validation error"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			}
		},
		"/api/logout": {
			"post": {
				"summary": "Logout user",
				"tags": ["Authentication"],
				"responses": {
					"204": {
						"description": "Logout successful"
					}
				}
			}
		},
		"/api/tasks": {
			"get": {
				"summary": "Get all tasks for authenticated user",
				"tags": ["Tasks"],
				"security": [{ "cookieAuth": [] }],
				"responses": {
					"200": {
						"description": "List of tasks",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"type": "object",
										"properties": {
											"_id": {
												"type": "string",
												"example": "507f1f77bcf86cd799439011"
											},
											"name": {
												"type": "string",
												"example": "Complete project"
											},
											"description": {
												"type": "string",
												"example": "Finish the MERN project"
											},
											"date": {
												"type": "string",
												"format": "date-time",
												"example": "2024-01-15T10:00:00Z"
											},
											"user": {
												"type": "string",
												"example": "507f1f77bcf86cd799439012"
											}
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "string",
											"example": "Unauthorized"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal server error"
					}
				}
			},
			"post": {
				"summary": "Create a new task",
				"tags": ["Tasks"],
				"security": [{ "cookieAuth": [] }],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"name": {
										"type": "string",
										"example": "New task"
									},
									"description": {
										"type": "string",
										"example": "Task description"
									},
									"date": {
										"type": "string",
										"format": "date-time",
										"example": "2024-01-20T10:00:00Z"
									}
								},
								"required": ["name"]
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Task created successfully",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"_id": {
											"type": "string",
											"example": "507f1f77bcf86cd799439011"
										},
										"name": {
											"type": "string",
											"example": "New task"
										},
										"description": {
											"type": "string",
											"example": "Task description"
										},
										"date": {
											"type": "string",
											"format": "date-time",
											"example": "2024-01-20T10:00:00Z"
										},
										"user": {
											"type": "string",
											"example": "507f1f77bcf86cd799439012"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"400": {
						"description": "Validation error"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			}
		},
		"/api/tasks/{id}": {
			"put": {
				"summary": "Update a task",
				"tags": ["Tasks"],
				"security": [{ "cookieAuth": [] }],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "507f1f77bcf86cd799439011"
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"name": {
										"type": "string",
										"example": "Updated task"
									},
									"description": {
										"type": "string",
										"example": "Updated description"
									},
									"date": {
										"type": "string",
										"format": "date-time",
										"example": "2024-01-25T10:00:00Z"
									}
								},
								"required": ["name"]
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Task updated successfully",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"_id": {
											"type": "string",
											"example": "507f1f77bcf86cd799439011"
										},
										"name": {
											"type": "string",
											"example": "Updated task"
										},
										"description": {
											"type": "string",
											"example": "Updated description"
										},
										"date": {
											"type": "string",
											"format": "date-time",
											"example": "2024-01-25T10:00:00Z"
										},
										"user": {
											"type": "string",
											"example": "507f1f77bcf86cd799439012"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Task not found",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "string",
											"example": "Task not found"
										}
									}
								}
							}
						}
					},
					"403": {
						"description": "Forbidden - can only edit own tasks",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "string",
											"example": "You can only edit your own tasks"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"400": {
						"description": "Validation error"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			},
			"delete": {
				"summary": "Delete a task",
				"tags": ["Tasks"],
				"security": [{ "cookieAuth": [] }],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "507f1f77bcf86cd799439011"
					}
				],
				"responses": {
					"204": {
						"description": "Task deleted successfully"
					},
					"404": {
						"description": "Task not found",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "string",
											"example": "Task not found"
										}
									}
								}
							}
						}
					},
					"403": {
						"description": "Forbidden - can only delete own tasks",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "string",
											"example": "You can only remove your own tasks"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			}
		}
	},
	"components": {
		"schemas": {
			"Task": {
				"type": "object",
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"date": {
						"type": "string",
						"format": "date-time"
					}
				}
			},
			"User": {
				"type": "object",
				"properties": {
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				}
			}
		}
	}
}